refactor: 공고 상세 페이지 수정
- 새로운 디자인 투두 아이템 적용
- 푸터 없앰
