--- src/app/(with-auth)/(with-footer)/job/[id]/page.module.css
@@ -1,349 +1,0 @@
-.main {
-  padding-top: 64px;
-  background-color: var(--n-100);
-}
-
-.container {
-  max-width: 1034px;
-  margin: 0 auto;
-  padding: 16px 16px 80px 16px;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 60px;
-  background-color: #ffffff;
-  padding: 0 16px;
-}
-
-.leftSection {
-  display: flex;
-  align-items: center;
-}
-
-.backButton {
-  /*background-color: var(--n-100);*/
-  background-color: white;
-  border-radius: 16px;
-  height: 48px;
-  width: 48px;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: background-color 0.2s;
-}
-
-.backButton:hover {
-  background-color: var(--n-200);
-}
-
-.rightSection {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.actionButton {
-  background-color: var(--primarymain-200);
-  border: 1px solid var(--primarymain-200);
-  border-radius: 4px;
-  height: 26px;
-  padding: 2px 8px;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.actionButton:hover {
-  background-color: var(--primarymain-300);
-}
-
-.actionButtonText {
-  color: var(--primaryjogak-1);
-  font-family: var(--label-1-reading-bold-font-family);
-  font-size: var(--label-1-reading-bold-font-size);
-  font-style: var(--label-1-reading-bold-font-style);
-  font-weight: var(--label-1-reading-bold-font-weight);
-  letter-spacing: var(--label-1-reading-bold-letter-spacing);
-  line-height: var(--label-1-reading-bold-line-height);
-  white-space: nowrap;
-}
-
-.iconButton {
-  background-color: white;
-  border-radius: 16px;
-  height: 48px;
-  width: 48px;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: background-color 0.2s;
-}
-
-.iconButton:hover {
-  background-color: var(--n-200);
-}
-
-/* Mobile responsive */
-@media (max-width: 768px) {
-  .container {
-    padding: 16px 16px 80px 16px;
-  }
-
-  .rightSection {
-    gap: 8px;
-  }
-
-  .progressTracker {
-    flex-direction: column;
-    gap: 16px;
-    align-items: stretch;
-  }
-
-  .progressContent {
-    width: 100%;
-    height: auto;
-  }
-
-  .notificationBtn {
-    width: 100%;
-  }
-
-  .jobDetails {
-    padding: 10px 16px;
-  }
-
-  .jobDetailsTop {
-    gap: 10px;
-  }
-
-  .jobDetailsLeft {
-    gap: 10px;
-  }
-
-  .jobDetailsBottom {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-  }
-}
-
-/* Jogak Categories */
-.jogakCategories {
-  display: flex;
-  gap: 16px;
-  margin-top: 12px;
-  width: 100%;
-  justify-content: center;
-}
-
-.jogakCategoryInstance {
-  width: 100%;
-  flex: 1;
-}
-
-@media (max-width: 768px) {
-  .jogakCategories {
-    flex-direction: column;
-    gap: 8px;
-    margin-top: 8px;
-  }
-}
-
-/* Memo Box */
-.memoBoxInstance {
-  margin-top: 12px;
-  width: 100%;
-  max-width: 1260px;
-}
-
-/* Dropdown menu styles */
-.moreMenu {
-  position: absolute;
-  top: 100%;
-  right: 0;
-  margin-top: 8px;
-  background-color: #ffffff;
-  border: 1px solid var(--n-300);
-  border-radius: 8px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-  min-width: 120px;
-  overflow: hidden;
-}
-
-.moreMenuItem {
-  display: block;
-  width: 100%;
-  padding: 12px 16px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  color: var(--n-800);
-  font-family: var(--body-2-normal-medium-font-family);
-  font-size: var(--body-2-normal-medium-font-size);
-  font-weight: var(--body-2-normal-medium-font-weight);
-  transition: background-color 0.2s;
-}
-
-.moreMenuItem.apply {
-  color: var(--primaryjogak-1);
-}
-
-.moreMenuItem.delete {
-  color: var(--red-3);
-}
-
-.moreMenuItem + .moreMenuItem {
-  border-top: 1px solid var(--n-200);
-}
-
-.moreMenuItem:hover {
-  background-color: var(--n-200);
-}
-
-.moreMenuItem:active {
-  background-color: var(--n-300);
-}
-
-/* Skeleton styles */
-.skeleton {
-  background: linear-gradient(
-    90deg,
-    var(--n-200) 25%,
-    var(--n-300) 50%,
-    var(--n-200) 75%
-  );
-  background-size: 200% 100%;
-  animation: loading 1.5s infinite;
-}
-
-@keyframes loading {
-  0% {
-    background-position: 200% 0;
-  }
-  100% {
-    background-position: -200% 0;
-  }
-}
-
-.skeleton--header button {
-  cursor: none;
-}
-
-.skeleton--header {
-  .actionButton {
-    background-color: var(--n-300);
-  }
-
-  .actionButtonText {
-    color: var(--n-600);
-  }
-}
-
-.skeleton--container {
-  max-width: 1066px;
-  margin: 0 auto;
-  padding: 16px;
-}
-
-.skeleton--jobDetails {
-  height: 82px;
-  width: 100%;
-  margin-bottom: 12px;
-  border-radius: 12px;
-}
-.skeleton--progressTracker {
-  height: 90px;
-  width: 100%;
-  margin-bottom: 12px;
-  border-radius: 12px;
-}
-.skeleton--jogakCategories {
-  height: 498px;
-  display: flex;
-  width: 100%;
-  gap: 16px;
-  margin-bottom: 12px;
-}
-.skeleton--jogakItem {
-  flex: 1;
-  height: 100%;
-  border-radius: 12px;
-}
-.skeleton--memobox {
-  height: 104px;
-  width: 100%;
-  border-radius: 12px;
-}
-
-@media screen and (max-width: 768px) {
-  .main {
-    padding-top: 56px;
-  }
-
-  .header {
-    height: 56px;
-    padding: 0 8px;
-  }
-
-  .leftSection {
-    gap: 8px;
-  }
-
-  .backButton {
-    width: 40px;
-    height: 40px;
-  }
-
-  .rightSection {
-    gap: 8px;
-  }
-
-  .actionButton {
-    height: 24px;
-  }
-
-  .iconButton {
-    height: 40px;
-    width: 40px;
-  }
-
-  .skeleton--container {
-    padding: 8px 16px;
-  }
-
-  .skeleton--jobDetails {
-    height: 82px;
-    margin-bottom: 8px;
-  }
-
-  .skeleton--progressTracker {
-    height: 138px;
-    margin-bottom: 8px;
-  }
-
-  .skeleton--jogakCategories {
-    flex-direction: column;
-    height: auto;
-    gap: 8px;
-  }
-
-  .skeleton--jogakItem {
-    height: 178px;
-    width: 100%;
-    flex: none;
-  }
-
-  .skeleton--memobox {
-    height: 110px;
-    width: 100%;
-    margin-bottom: 121px;
-  }
-}
