+++ src/app/(with-auth)/job/[id]/page.module.css
@@ -1,0 +1,74 @@
+.main {
+  position: relative;
+  box-sizing: border-box;
+  top: 64px; /* 헤더 높이만큼 내려줌 */
+  min-height: calc(100vh - 64px);
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: var(--n-100);
+  padding: 0px 20px 60px 20px;
+}
+
+.content {
+  max-width: 1560px;
+  width: 100%;
+}
+
+.todoListSection {
+  margin-top: 17px;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(343px, 1fr));
+  justify-items: center;
+  gap: 20px;
+}
+
+.checkedConfirmation {
+  position: fixed;
+  bottom: 22px;
+  left: 50%;
+  transform: translateX(-50%);
+  max-width: 770px;
+  /* min-width: 343px; */
+  width: 100%;
+  height: 80px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  border-radius: 12px;
+  border: 1px solid var(--primarymain-200);
+  padding: 0 16px;
+  background-color: #ffffff;
+  box-shadow: 0px 4px 12px 0 rgba(0, 0, 0, 0.1);
+}
+
+.checkedText {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding-left: 8px;
+  font-family: var(--body-1-normal-bold-font-family);
+  font-size: var(--body-1-normal-bold-font-size);
+  font-style: var(--body-1-normal-bold-font-style);
+  font-weight: var(--body-1-normal-bold-font-weight);
+  letter-spacing: var(--body-1-normal-bold-letter-spacing);
+  color: var(--n-900);
+}
+
+.checkedText span {
+  font-family: var(--body-1-normal-regular-font-family);
+  font-size: var(--body-1-normal-regular-font-size);
+  font-style: var(--body-1-normal-regular-font-style);
+  font-weight: var(--body-1-normal-regular-font-weight);
+  letter-spacing: var(--body-1-normal-regular-letter-spacing);
+  color: var(--primaryjogak-1);
+}
+
+@media screen and (max-width: 576px) {
+  .todoListSection {
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 20px;
+  }
+}
